backend/requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
bokeh==3.6.2
cffi==1.17.1
chardet==5.2.0
click==8.1.7
cloudpickle==3.1.1
contourpy==1.3.1
cryptography==44.0.0
dask==2024.12.1
dask-expr==1.1.21
distributed==2024.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.5
locket==1.0.0
lz4==4.3.3
Mako==1.3.8
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.1.2
packaging==24.2
pandas==2.2.3
partd==1.4.2
passlib==1.7.4
pillow==11.1.0
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
starlette==0.41.0
tblib==3.0.0
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
xyzservices==2025.1.0
zict==3.0.0
